requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
appdirs==1.4.4            # via black
beautifulsoup4==4.9.3     # via -r requirements.in
black==20.8b1             # via -r requirements.in
certifi==2020.12.5        # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via black, flask, pip-tools
flask-api==2.0            # via -r requirements.in
flask-sqlalchemy==2.4.4   # via -r requirements.in
flask==1.1.2              # via flask-api, flask-sqlalchemy
idna==2.10                # via requests
isort==5.6.4              # via -r requirements.in
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
markupsafe==1.1.1         # via jinja2
mypy-extensions==0.4.3    # via black
pathspec==0.8.1           # via black
pika==1.1.0               # via -r requirements.in
pip-tools==5.4.0          # via -r requirements.in
pyodbc==4.0.30            # via sqlalchemy-pyodbc-mssql
python-dateutil==2.8.1    # via -r requirements.in
python-dotenv==0.15.0     # via -r requirements.in
python-http-client==3.3.1  # via sendgrid
regex==2020.11.13         # via black
requests==2.25.0          # via -r requirements.in
sendgrid==6.4.8           # via -r requirements.in
six==1.15.0               # via pip-tools, python-dateutil
soupsieve==2.1            # via beautifulsoup4
sqlalchemy-pyodbc-mssql==0.1.0  # via -r requirements.in
sqlalchemy==1.3.22        # via flask-sqlalchemy, sqlalchemy-pyodbc-mssql
starkbank-ecdsa==1.1.0    # via sendgrid
toml==0.10.2              # via black
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via black
urllib3==1.26.2           # via requests
werkzeug==1.0.1           # via flask

# The following packages are considered to be unsafe in a requirements file:
# pip
